move response records to its own file, add util class to map, update controller to use util class to map bean
